Fix auto-negotiation by actually enabling it.
Both the upstream Linux ax88179_178a driver and the official ASIX driver don't enable auto-negotiation for my AX88179A, resulting in 10 Mbps half-duplex speeds.
I need to manually run `mii-tool --restart eth1` to get auto-negotiation.

This patch turns on auto-negotiation for AX88179A by setting `BMCR_ANRESTART | BMCR_ANENABLE`.
This is already done for the 2.5Gbps AX88279 chip, but the code path was not enabled for the AX88179A.

This fixes the problem for us, but:
- I don't know what's best practice on setting `BMCR_ANENABLE`.
  Many upstream Linux drivers do it, others don't.
- I can't say if other nearby commands from the AX88279 code path are also relevant to us.
- I don't understand why nobody before us has hit this problem.